requirements.txt
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
matplotlib>=3.7.0 